requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
bleach==3.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-results==2.0.1
django-filter==2.4.0
django-memcached==0.1.2
djangorestframework==3.12.2
drf-api-logger==1.0.2
gunicorn==20.0.4
kombu==5.0.2
Markdown==3.3.3
packaging==20.9
prompt-toolkit==3.0.16
pyparsing==2.4.7
python-memcached==1.59
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5